detect: make progress values uint8_t
Also reflect that it can be >= 48, like in prefilter.

+++ src/detect-engine-prefilter.h
@@ -40,8 +40,8 @@ typedef struct DetectTransaction_ {
     /* original value to track changes. */
     const uint8_t detect_progress_orig;
 
-    const int tx_progress;
-    const int tx_end_state;
+    const uint8_t tx_progress;
+    const uint8_t tx_end_state;
 } DetectTransaction;
 
 typedef struct PrefilterStore_ {
